motzip-server/pyproject.toml
[project]
name = "motzip-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.12",
    "httpx==0.28.1",
    "pydantic==2.11.3",
    "python-multipart==0.0.20",
    "uvicorn==0.34.2",
    "twilio>=9.0.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.73.1",
]
